Documentation, refactor GET route (function sent to controller), comment code
