+++ api/Rutas/GetDietsRoutines.md
@@ -1,0 +1,20 @@
+##### Objective:
+
+- Get all the user's diets and routines
+- id user by query
+
+##### Route:
+
+- http://localhost:8200/api/userDietsRoutines
+
+##### Headers:
+
+- Token (en minusculas)
+
+##### Method:
+
+- GET
+
+##### File:
+
+- User_diets_routines.js
