+++ api/Rutas/GetUsersBoughtItem.md
@@ -1,0 +1,21 @@
+##### Objective:
+
+- Get all the users who bought an specific item.
+- id product by query.
+- id must be verify in existence before sent out.
+
+##### Route:
+
+- http://localhost:8200/api/soldItems
+
+##### Headers:
+
+- Token (en minusculas)
+
+##### Method:
+
+- GET
+
+##### File:
+
+- User_diets_routines.js
